refactor: AdminEditor 파일 분리
